ブロックにフォーカスが当たっててもプレビューモードの時は開かない

+++ moorestech_client/Assets/Scripts/Client.Network/API/VanillaApiSendOnly.cs
@@ -31,7 +31,7 @@ namespace Client.Network.API
             _packetSender.Send(request);
         }
 
-        public void PlaceHotBarBlock(Vector3Int pos, short hotBarSlot, BlockDirection blockDirection)
+        public void PlaceHotBarBlock(Vector3Int pos, int hotBarSlot, BlockDirection blockDirection)
         {
             var request = new SendPlaceHotBarBlockProtocolMessagePack(_playerId, blockDirection, hotBarSlot, pos);
             _packetSender.Send(request);
